move collector helper functions and tests to the provider package, use the sdk configured by provider.go
