+++ provider/resource_collector.go
@@ -3,11 +3,16 @@ package provider
 import (
 	"github.com/pkg/errors"
 	"strings"
+	"time"
 
 	"github.com/BlueMedoraPublic/terraform-provider-bindplane/provider/bindplane/collector"
 
 	"github.com/hashicorp/terraform/helper/schema"
+	"github.com/BlueMedoraPublic/bpcli/bindplane/sdk"
 )
+
+// Max time to wait for a collector to deploy in seconds
+const collectorTimeout = int64(300)
 
 func resourceCollector() *schema.Resource {
 	return &schema.Resource{
@@ -39,7 +44,7 @@ with a userdata script for collector deployment
 */
 func resourceCollectorCreate(d *schema.ResourceData, m interface{}) error {
 	name := d.Get("name").(string)
-	id, err := collector.WaitForAPI(name)
+	id, err := collector.waitForAPI(bp, name)
 	if err != nil {
 		return errors.Wrap(err, "terraform bindplane_collector resource failed while waiting for collector to appear in bindplane api")
 	}
@@ -51,7 +56,7 @@ func resourceCollectorCreate(d *schema.ResourceData, m interface{}) error {
 
 // resourceCollectorRead checks to see if a specific collector exists
 func resourceCollectorRead(d *schema.ResourceData, m interface{}) error {
-	if err := collector.Read(d.Id()); err != nil {
+	if _, err := bp.GetCollector(d.Id()); err != nil {
 		/*
 			It is possible the collector in the tf state does not exist.
 			If this happens, remove it from the tf state by setting
@@ -73,9 +78,39 @@ the create resource does not create a virtual machine
 */
 func resourceCollectorDelete(d *schema.ResourceData, m interface{}) error {
 	id := d.Get("group").(string) + "/" + d.Id()
-	if err := collector.Delete(id); err != nil {
+	if err := bp.DeleteCollector(id); err != nil {
 		return err
 	}
 	d.SetId("")
 	return nil
+}
+
+
+/*
+WaitForAPI checks the BindPlane API every ten seconds
+until a collector with the name 'collectorName' is found. The
+collector id is returned.
+*/
+func waitForAPI(bp *sdk.BindPlane, collectorName string) (string, error) {
+	startTime := time.Now().Unix()
+
+	for {
+		collectors, err := bp.GetCollectors()
+		if err != nil {
+			return "", errors.Wrap(err, "sdk.GetCollectors() returned an error")
+		}
+
+		for _, collector := range collectors {
+			if collector.Name == collectorName {
+				return collector.ID, nil
+			}
+		}
+
+		time.Sleep(10 * time.Second)
+		timeCurrent := time.Now().Unix()
+		if (timeCurrent - startTime) > collectorTimeout {
+			msg := "Timeout exceeded for collector creation: " + "CollectorName: " + collectorName
+			return "", errors.New(msg)
+		}
+	}
 }
